+++ commands_extra/cs_variables.sh
@@ -15,6 +15,7 @@ RED=$(printf '\033[31m')
 GREEN=$(printf '\033[32m')
 YELLOW=$(printf '\033[33m')
 BLUE=$(printf '\033[34m')
+PURPLE=$(printf '\033[35m')
 
 BOLD=$(printf '\033[1m')
 UNDER=$(printf '\033[4m')