+++ commands_extra/ros2/shell/install_ros2_common.sh
@@ -94,16 +94,16 @@ ros-$ros2_distro-tf2-tools \
 ros-$ros2_distro-tf-transformations
 
 #install ros2 bag deps
-if [ $ros2_arch == "amd" ]; then 
-    sudo apt install -y \
-    libroscpp-serialization0d \
-    ros-$ros2_distro-nmea-msgs \
-    ros-$ros2_distro-rosbag2  \
-    ros-$ros2_distro-rosbag2-storage \
-    ros-$ros2_distro-rosbag2-storage-default-plugins \
-    ros-$ros2_distro-ros2bag \
-    ros-$ros2_distro-rosbag2-transport 
-fi
+# if [ $ros2_arch == "amd" ]; then 
+#     sudo apt install -y \
+#     libroscpp-serialization0d \
+#     ros-$ros2_distro-nmea-msgs \
+#     ros-$ros2_distro-rosbag2  \
+#     ros-$ros2_distro-rosbag2-storage \
+#     ros-$ros2_distro-rosbag2-storage-default-plugins \
+#     ros-$ros2_distro-ros2bag \
+#     ros-$ros2_distro-rosbag2-transport 
+# fi
 
 #add ros to bashrc
 sudo echo "source /opt/ros/$ros2_distro/setup.bash" >> ~/.bashrc