+++ CHANGELOG.rst
@@ -3,8 +3,10 @@
 2.0.2 (2023-09-08)
 
 - 增加docker支持， 进入commands_docker运行`docker build -t rcm:latest .` 生成docker镜像
+- 镜像已经上传到docker-hub,可以直接使用运行 docker run -it ncnynl/rcm:latest 
 - 通过命令`docker run  -it rcm:latest` 进入终端
 - 增加备份私有目录和私有脚本的同步脚本，同步后目录和文件位于~/tools/commands_private/
+- 增加脚本版本状态字段：DEV 开发版本，BETA 预览版本 ，RELEASE 正式版. 生成的脚本默认是DEV状态
 
 
 2.0.1 (2023-09-07)